update rfProtocol data type

+++ radio/src/datastructs.h
@@ -412,7 +412,7 @@ PACK(struct TelemetrySensor {
 
 PACK(struct ModuleData {
   uint8_t type; // : 4; // use full byte to reduce code size on PCBI6X
-  int8_t rfProtocol; //  : 4
+  uint8_t rfProtocol; //  : 4
   uint8_t channelsStart;
   int8_t channelsCount;      // 0=8 channels
   uint8_t failsafeMode : 4;  // only 3 bits used
@@ -459,7 +459,7 @@ PACK(struct ModuleData {
   })
 
   NOBACKUP(inline void setMultiProtocol(uint8_t proto) {
-    rfProtocol = (int8_t)(proto);
+    rfProtocol = proto;
   })
 });
